Break out Search Source base class
